Minor changes due to comments on pull-request
