+++ jgrapht-core/src/main/java/org/jgrapht/alg/EulerianCircuit.java
@@ -21,6 +21,7 @@ import java.util.*;
 
 import org.jgrapht.*;
 import org.jgrapht.graph.*;
+import org.jgrapht.util.*;
 
 /**
  * This algorithm will check whether a graph is Eulerian (hence it contains an
@@ -33,18 +34,62 @@ import org.jgrapht.graph.*;
  */
 public abstract class EulerianCircuit
 {
+
     /**
-     * This method will check whether the graph passed in is Eulerian or not.
+     * Test whether a graph is Eulerian. An undirected graph is Eulerian if it is connected and each
+     * vertex has an even degree.
      *
-     * @param g The graph to be checked
+     * @param graph the input graph
      * @param <V> the graph vertex type
      * @param <E> the graph edge type
      *
-     * @return true for Eulerian and false for non-Eulerian
+     * @return true if the graph is Eulerian, false otherwise
+     * @deprecated in favor of {@link #isEulerian(Graph)}
      */
-    public static <V, E> boolean isEulerian(UndirectedGraph<V, E> g)
+    @Deprecated
+    public static <V, E> boolean isEulerian(UndirectedGraph<V, E> graph)
     {
-        return GraphTests.isEulerian(g);
+        return isEulerian((Graph<V, E>) graph);
+    }
+
+    /**
+     * Test whether a graph is Eulerian. An undirected graph is Eulerian if it is connected and each
+     * vertex has an even degree. A directed graph is Eulerian if it is strongly connected and each
+     * vertex has the same incoming and outgoing degree.
+     *
+     * @param graph the input graph
+     * @param <V> the graph vertex type
+     * @param <E> the graph edge type
+     *
+     * @return true if the graph is Eulerian, false otherwise
+     */
+    public static <V, E> boolean isEulerian(Graph<V, E> graph)
+    {
+        Objects.requireNonNull(graph, "Graph cannot be null");
+        if (graph instanceof DirectedGraph) {
+            DirectedGraph<V, E> dg = TypeUtil.uncheckedCast(graph, null);
+            for (V v : dg.vertexSet()) {
+                if (dg.inDegreeOf(v) != dg.outDegreeOf(v)) {
+                    return false;
+                }
+            }
+            if (!GraphTests.isStronglyConnected(dg)) {
+                return false;
+            }
+        } else if (graph instanceof UndirectedGraph) {
+            UndirectedGraph<V, E> ug = TypeUtil.uncheckedCast(graph, null);
+            for (V v : ug.vertexSet()) {
+                if (ug.degreeOf(v) % 2 == 1) {
+                    return false;
+                }
+            }
+            if (!GraphTests.isConnected(ug)) {
+                return false;
+            }
+        } else {
+            throw new IllegalArgumentException("Graph must be directed or undirected");
+        }
+        return true;
     }
 
     /**
