+++ jgrapht-core/src/test/java/org/jgrapht/generate/GnpRandomBipartiteGraphGeneratorTest.java
@@ -38,7 +38,7 @@ public class GnpRandomBipartiteGraphGeneratorTest
         GraphGenerator<Integer, DefaultEdge, Integer> gen =
             new GnpRandomBipartiteGraphGenerator<>(0, 0, 0.5);
         DirectedGraph<Integer, DefaultEdge> g = new DirectedPseudograph<>(DefaultEdge.class);
-        gen.generateGraph(g, new IntegerVertexFactory(), null);
+        gen.generateGraph(g, new IntegerVertexFactory(1), null);
         assertEquals(0, g.vertexSet().size());
         assertEquals(0, g.edgeSet().size());
     }
@@ -75,7 +75,7 @@ public class GnpRandomBipartiteGraphGeneratorTest
         GraphGenerator<Integer, DefaultEdge, Integer> gen =
             new GnpRandomBipartiteGraphGenerator<>(4, 4, 0.5, SEED);
         DirectedGraph<Integer, DefaultEdge> g = new DirectedPseudograph<>(DefaultEdge.class);
-        gen.generateGraph(g, new IntegerVertexFactory(), null);
+        gen.generateGraph(g, new IntegerVertexFactory(1), null);
 
         int[][] edges = { { 5, 1 }, { 1, 6 }, { 6, 1 }, { 1, 7 }, { 1, 8 }, { 2, 5 }, { 6, 2 },
             { 7, 2 }, { 2, 8 }, { 5, 3 }, { 3, 6 }, { 7, 3 }, { 3, 8 }, { 4, 5 }, { 5, 4 },
@@ -93,7 +93,7 @@ public class GnpRandomBipartiteGraphGeneratorTest
         GraphGenerator<Integer, DefaultEdge, Integer> gen =
             new GnpRandomBipartiteGraphGenerator<>(4, 4, 1.0, SEED);
         DirectedGraph<Integer, DefaultEdge> g = new DirectedPseudograph<>(DefaultEdge.class);
-        gen.generateGraph(g, new IntegerVertexFactory(), null);
+        gen.generateGraph(g, new IntegerVertexFactory(1), null);
 
         assertEquals(4 + 4, g.vertexSet().size());
         assertEquals(32, g.edgeSet().size());
@@ -104,7 +104,7 @@ public class GnpRandomBipartiteGraphGeneratorTest
         GraphGenerator<Integer, DefaultEdge, Integer> gen =
             new GnpRandomBipartiteGraphGenerator<>(4, 4, 0.1, SEED);
         DirectedGraph<Integer, DefaultEdge> g = new DirectedPseudograph<>(DefaultEdge.class);
-        gen.generateGraph(g, new IntegerVertexFactory(), null);
+        gen.generateGraph(g, new IntegerVertexFactory(1), null);
 
         int[][] edges = { { 5, 1 }, { 7, 3 }, { 3, 8 }, { 8, 4 } };
 
@@ -120,7 +120,7 @@ public class GnpRandomBipartiteGraphGeneratorTest
         GraphGenerator<Integer, DefaultEdge, Integer> gen =
             new GnpRandomBipartiteGraphGenerator<>(4, 4, 0.5, SEED);
         UndirectedGraph<Integer, DefaultEdge> g = new SimpleGraph<>(DefaultEdge.class);
-        gen.generateGraph(g, new IntegerVertexFactory(), null);
+        gen.generateGraph(g, new IntegerVertexFactory(1), null);
 
         int[][] edges = { { 1, 6 }, { 1, 7 }, { 1, 8 }, { 2, 5 }, { 2, 7 }, { 3, 5 }, { 3, 8 },
             { 4, 6 }, { 4, 7 } };
@@ -137,7 +137,7 @@ public class GnpRandomBipartiteGraphGeneratorTest
         GraphGenerator<Integer, DefaultEdge, Integer> gen =
             new GnpRandomBipartiteGraphGenerator<>(4, 4, 1.0, SEED);
         UndirectedGraph<Integer, DefaultEdge> g = new SimpleGraph<>(DefaultEdge.class);
-        gen.generateGraph(g, new IntegerVertexFactory(), null);
+        gen.generateGraph(g, new IntegerVertexFactory(1), null);
 
         assertEquals(4 + 4, g.vertexSet().size());
         assertEquals(4 * 4, g.edgeSet().size());
@@ -148,23 +148,11 @@ public class GnpRandomBipartiteGraphGeneratorTest
         GraphGenerator<Integer, DefaultEdge, Integer> gen =
             new GnpRandomBipartiteGraphGenerator<>(4, 4, 0.0, SEED);
         UndirectedGraph<Integer, DefaultEdge> g = new SimpleGraph<>(DefaultEdge.class);
-        gen.generateGraph(g, new IntegerVertexFactory(), null);
+        gen.generateGraph(g, new IntegerVertexFactory(1), null);
 
         assertEquals(4 + 4, g.vertexSet().size());
         assertEquals(0, g.edgeSet().size());
     }
-
-    private class IntegerVertexFactory
-        implements VertexFactory<Integer>
-    {
-        private int id;
-
-        @Override
-        public Integer createVertex()
-        {
-            return ++id;
-        }
-    };
 
 }
